nass-737: optimize recursion
